campie/pyproject.toml
[tool.poetry]
name = "campie"
version = "0.3.0"
description = "Python APIs to simulate various CAMs on GPUs at scale"
readme = "README.md"
authors = []
packages = [{ include = "campie" }]

[tool.poetry.dependencies]
python = "^3.8"

cupy-cuda110 = { version = "^12.1.0", optional = true }
cupy-cuda111 = { version = "^12.1.0", optional = true }
cupy-cuda11x = { version = "^12.1.0", optional = true }
cupy-cuda12x = { version = "^12.1.0", optional = true }

numpy = "^1.22"

[tool.poetry.extras]
cu110 = ["cupy-cuda110"]
cu111 = ["cupy-cuda111"]
cu11x = ["cupy-cuda11x"]
cu12x = ["cupy-cuda12x"]

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
ruff = "^0.0.278"
ipykernel = "^6.25.1"

[tool.poetry.scripts]
gen-reference = "scripts.gen_reference:main"
format = "scripts.format:main"

[tool.isort]
profile = "black"
